(#4095) - 'destroyed' event emits on all DBs
