+++ lib/constructor.js
@@ -14,6 +14,37 @@ function defaultCallback(err) {
   }
 }
 
+// OK, so here's the deal. Consider this code:
+//     var db1 = new PouchDB('foo');
+//     var db2 = new PouchDB('foo');
+//     db1.destroy();
+// ^ these two both need to emit 'destroyed' events,
+// as well as the PouchDB constructor itself.
+// So we have one db object (whichever one got destroy() called on it)
+// responsible for emitting the initial event, which then gets emitted
+// by the constructor, which then broadcasts it to any other dbs
+// that may have been created with the same name.
+function prepareForDestruction(self, opts) {
+
+  function constructorDestructionListener(name) {
+    if (name === opts.originalName) {
+      self.removeListener('destroyed', destructionListener);
+      self.emit('destroyed', self);
+    }
+  }
+
+  function destructionListener() {
+    // we destroyed ourselves, so no need to listen on the constructor
+    PouchDB.removeListener('destroyed', constructorDestructionListener);
+    PouchDB.emit('destroyed', opts.originalName);
+    //so we don't have to sift through all dbnames
+    PouchDB.emit(opts.originalName, 'destroyed');
+  }
+
+  PouchDB.once('destroyed', constructorDestructionListener);
+  self.once('destroyed', destructionListener);
+}
+
 utils.inherits(PouchDB, Adapter);
 function PouchDB(name, opts, callback) {
 
@@ -127,13 +158,8 @@ function PouchDB(name, opts, callback) {
         }
         return;
       }
-      function destructionListener() {
-        PouchDB.emit('destroyed', opts.originalName);
-        //so we don't have to sift through all dbnames
-        PouchDB.emit(opts.originalName, 'destroyed');
-        self.removeListener('destroyed', destructionListener);
-      }
-      self.on('destroyed', destructionListener);
+      prepareForDestruction(self, opts);
+
       self.emit('created', self);
       PouchDB.emit('created', opts.originalName);
       self.taskqueue.ready(self);
