+++ lib/mapreduce/create-view.js
@@ -65,7 +65,7 @@ module.exports = function (opts) {
           if (!temporary) {
             sourceDB._cachedViews = sourceDB._cachedViews || {};
             sourceDB._cachedViews[viewSignature] = view;
-            view.db.on('destroyed', function () {
+            view.db.once('destroyed', function () {
               delete sourceDB._cachedViews[viewSignature];
             });
           }
