+++ lib/setup.js
@@ -2,7 +2,7 @@
 
 var PouchDB = require("./constructor");
 var utils = require('./utils');
-var EventEmitter = require('events').EventEmitter;
+var EE = require('events').EventEmitter;
 var hasLocalStorage = require('./deps/env/hasLocalStorage');
 
 PouchDB.adapters = {};
@@ -10,23 +10,18 @@ PouchDB.preferredAdapters = [];
 
 PouchDB.prefix = '_pouch_';
 
-var eventEmitter = new EventEmitter();
+var eventEmitter = new EE();
 
-var eventEmitterMethods = [
-  'on',
-  'addListener',
-  'emit',
-  'listeners',
-  'once',
-  'removeAllListeners',
-  'removeListener',
-  'setMaxListeners'
-];
+function bindEventEmitterMethods(Pouch) {
+  Object.keys(EE.prototype).forEach(function (key) {
+    if (typeof EE.prototype[key] === 'function') {
+      Pouch[key] = eventEmitter[key].bind(eventEmitter);
+    }
+  });
+}
 
-eventEmitterMethods.forEach(function (method) {
-  PouchDB[method] = eventEmitter[method].bind(eventEmitter);
-});
-PouchDB.setMaxListeners(0);
+bindEventEmitterMethods(PouchDB);
+
 PouchDB.parseAdapter = function (name, opts) {
   var match = name.match(/([a-z\-]*):\/\/(.*)/);
   var adapter, adapterName;
@@ -146,10 +141,7 @@ PouchDB.defaults = function (defaultOpts) {
     return PouchDB.destroy(name, opts, callback);
   });
 
-  eventEmitterMethods.forEach(function (method) {
-    PouchAlt[method] = eventEmitter[method].bind(eventEmitter);
-  });
-  PouchAlt.setMaxListeners(10);
+  bindEventEmitterMethods(PouchAlt);
 
   PouchAlt.preferredAdapters = PouchDB.preferredAdapters.slice();
   Object.keys(PouchDB).forEach(function (key) {
