+++ tests/mapreduce/index.html
@@ -25,6 +25,7 @@
 </script>
 <script src='../integration/utils.js'></script>
 <script src="test.mapreduce.js"></script>
+<script src="test.persisted.js"></script>
 <script src='test.views.js'></script>
 <script type="text/javascript" src="../integration/webrunner.js"></script>
 </body>
